ai-teacher-backend/requirements.txt
# FastAPI and ASGI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Data validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# HTTP client
httpx>=0.26.0

# LLM
zhipuai>=2.0.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Utilities
python-dateutil>=2.8.2
uuid6>=2024.1.12

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0

# Development
black>=24.1.0
ruff>=0.1.0
mypy>=1.8.0
